+++ src/Mirage.Godot.Example1/Scripts/PlayerController.cs
@@ -64,7 +64,7 @@ namespace Example1
             direction.Y = 0;
             direction = direction.Normalized();
             var angle = direction.SignedAngleTo(Vector3.Forward, Vector3.Down);
-            return angle;
+            return (float)angle;
         }
 
         private static Vector3 GetInput()
