+++ src/Mirage.Godot/Scripts/Serialization/Packers/QuaternionPacker.cs
@@ -177,24 +177,24 @@ namespace Mirage.Serialization
             switch (largestIndex)
             {
                 case 0:
-                    a = quaternion.Y;
-                    b = quaternion.Z;
-                    c = quaternion.W;
+                    a = (float)quaternion.Y;
+                    b = (float)quaternion.Z;
+                    c = (float)quaternion.W;
                     return;
                 case 1:
-                    a = quaternion.X;
-                    b = quaternion.Z;
-                    c = quaternion.W;
+                    a = (float)quaternion.X;
+                    b = (float)quaternion.Z;
+                    c = (float)quaternion.W;
                     return;
                 case 2:
-                    a = quaternion.X;
-                    b = quaternion.Y;
-                    c = quaternion.W;
+                    a = (float)quaternion.X;
+                    b = (float)quaternion.Y;
+                    c = (float)quaternion.W;
                     return;
                 case 3:
-                    a = quaternion.X;
-                    b = quaternion.Y;
-                    c = quaternion.Z;
+                    a = (float)quaternion.X;
+                    b = (float)quaternion.Y;
+                    c = (float)quaternion.Z;
                     return;
                 default:
                     ThrowIfOutOfRange();
